requirements.txt
pandas==1.3.4
numpy==1.21.3
yfinance==0.1.64
yahoo_fin==0.8.9.1
tqdm==4.62.3
scikit-learn==1.0
statsmodels==0.13.0
xgboost==1.5.0
matplotlib==3.4.3
seaborn==0.11.2
giotto-tda==0.5.1
PyYAML==6.0
mlflow==1.20.2
luigi==3.0.3
streamlit==1.1.0